git-svn-id: https://sfs.atcorau.local/svn/SoftwareRepository/trunk/PN022@237 a8a9b14d-6b92-5b4b-9153-4cd90bb22c97

+++ Scor/TestDal/DalPwvDataStructTest.cpp
@@ -94,13 +94,13 @@ namespace TestDal {
 			void DalPwvDataStructConstructorTest()
 			{
 				DalPwvDataStruct^  target = (gcnew DalPwvDataStruct());
-				short expectedValue = -1;
+				unsigned long expectedValue = -1;
 
 				Assert::AreEqual(target->tonometerData, expectedValue);
 				Assert::AreEqual(target->cuffPulseData, expectedValue);
 				Assert::AreEqual(target->cuffPressure, expectedValue);
 				Assert::AreEqual(target->countdownTimer, expectedValue);
- 			}
+			}
 	};
 }
 namespace TestDal {
